# docs/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "phpList4 REST API",
        "version": "4.0.0"
    },
    "paths": {
...
        "/api/v2/lists/{list}": {
            "get": {
                "tags": [
                    "lists"
                ],
                "summary": "Gets a subscriber list.",
                "description": "Returns a single subscriber lists with specified ID",
                "operationId": "PhpList\\RestBundle\\Controller\\ListController::getAction",
                "parameters": [
                    {
                        "name": "list",
                        "in": "path",
                        "description": "List ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID obtained from authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": [
                                    {
                                        "name": "News",
                                        "description": "News (and some fun stuff)",
                                        "creation_date": "2016-06-22T15:01:17+00:00",
                                        "list_position": 12,
                                        "subject_prefix": "phpList",
                                        "public": true,
                                        "category": "news",
                                        "id": 1
                                    }
                                ]
                            }
                        }
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Gets a subscriber list.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Gets a subscriber list.",
                                            "type": "string",
                                            "example": "There is no list with that ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "lists"
                ],
                "summary": "Deletes a list.",
                "description": "Deletes a single subscriber list passed as",
                "operationId": "PhpList\\RestBundle\\Controller\\ListController::deleteAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "header",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a subscriber list.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password or You do not have access to this session."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a subscriber list.",
                                            "type": "string",
                                            "example": "There is no session with that ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v2/sessions": {
            "post": {
                "tags": [
                    "sessions"
                ],
                "summary": "Log in or create new session.",
                "description": "Given valid login data, this will generate a login token that will be valid for 1 hour",
                "operationId": "PhpList\\RestBundle\\Controller\\SessionController::postAction",
                "requestBody": {
                    "description": "Pass session credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login_name",
                                    "password"
                                ],
                                "properties": {
                                    "login_name": {
                                        "description": "Creates a new session (if the provided credentials are valid).",
                                        "type": "string",
                                        "format": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "description": "Creates a new session (if the provided credentials are valid).",
                                        "type": "string",
                                        "format": "password",
                                        "example": "eetIc/Gropvoc1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "integer",
                                            "example": "1234"
                                        },
                                        "key": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "2cfe100561473c6cdd99c9e2f26fa974"
                                        },
                                        "expiry": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "2017-07-20T18:22:48+00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "Empty json, invalid data and or incomplete data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Creates a new session (if the provided credentials are valid).",
                                            "type": "string",
                                            "example": "Not authorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v2/sessions/{session}": {
            "delete": {
                "tags": [
                    "sessions"
                ],
                "summary": "Delete a session.",
                "description": "Delete the session passed as paramater",
                "operationId": "PhpList\\RestBundle\\Controller\\SessionController::deleteAction",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a session.",
                                            "type": "string",
                                            "example": "No valid session key was provided as basic auth password or You do not have access to this session."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Deletes a session.",
                                            "type": "string",
                                            "example": "There is no session with that ID."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}